move `const_eval` and `check_match` out of `librustc`
