+++ src/librustc/lib.rs
@@ -91,8 +91,7 @@ pub mod lint;
 pub mod middle {
     pub mod astconv_util;
     pub mod expr_use_visitor; // STAGE0: increase glitch immunity
-    pub mod check_match;
-    pub mod const_eval;
+    pub mod const_val;
     pub mod const_qualif;
     pub mod cstore;
     pub mod dataflow;
