+++ src/librustc_const_eval/Cargo.toml
@@ -1,16 +1,19 @@
 [package]
 authors = ["The Rust Project Developers"]
-name = "rustc_passes"
+name = "rustc_const_eval"
 version = "0.0.0"
 
 [lib]
-name = "rustc_passes"
+name = "rustc_const_eval"
 path = "lib.rs"
 crate-type = ["dylib"]
 
 [dependencies]
 log = { path = "../liblog" }
+serialize = { path = "../libserialize" }
 rustc = { path = "../librustc" }
-rustc_const_eval = { path = "../librustc_const_eval" }
 rustc_front = { path = "../librustc_front" }
+rustc_back = { path = "../librustc_back" }
+rustc_const_math = { path = "../librustc_const_math" }
 syntax = { path = "../libsyntax" }
+graphviz = { path = "../libgraphviz" }
