+++ src/librustc_const_eval/check_match.rs
@@ -1,4 +1,4 @@
-// Copyright 2012-2014 The Rust Project Developers. See the COPYRIGHT
+// Copyright 2012-2016 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
 //
@@ -12,22 +12,22 @@ pub use self::Constructor::*;
 use self::Usefulness::*;
 use self::WitnessPreference::*;
 
-use dep_graph::DepNode;
-use middle::const_eval::{compare_const_vals, ConstVal};
-use middle::const_eval::{eval_const_expr, eval_const_expr_partial};
-use middle::const_eval::{const_expr_to_pat, lookup_const_by_id};
-use middle::const_eval::EvalHint::ExprTypeChecked;
-use middle::def::*;
-use middle::def_id::{DefId};
-use middle::expr_use_visitor::{ConsumeMode, Delegate, ExprUseVisitor};
-use middle::expr_use_visitor::{LoanCause, MutateMode};
-use middle::expr_use_visitor as euv;
-use infer;
-use middle::mem_categorization::{cmt};
-use middle::pat_util::*;
-use traits::ProjectionMode;
-use ty::*;
-use ty;
+use rustc::dep_graph::DepNode;
+use rustc::middle::const_val::ConstVal;
+use ::{eval_const_expr, eval_const_expr_partial, compare_const_vals};
+use ::{const_expr_to_pat, lookup_const_by_id};
+use ::EvalHint::ExprTypeChecked;
+use rustc::middle::def::*;
+use rustc::middle::def_id::{DefId};
+use rustc::middle::expr_use_visitor::{ConsumeMode, Delegate, ExprUseVisitor};
+use rustc::middle::expr_use_visitor::{LoanCause, MutateMode};
+use rustc::middle::expr_use_visitor as euv;
+use rustc::infer;
+use rustc::middle::mem_categorization::{cmt};
+use rustc::middle::pat_util::*;
+use rustc::traits::ProjectionMode;
+use rustc::ty::*;
+use rustc::ty;
 use std::cmp::Ordering;
 use std::fmt;
 use std::iter::{FromIterator, IntoIterator, repeat};
@@ -44,7 +44,7 @@ use syntax::codemap::{Span, Spanned, DUMMY_SP};
 use rustc_front::fold::{Folder, noop_fold_pat};
 use rustc_front::print::pprust::pat_to_string;
 use syntax::ptr::P;
-use util::nodemap::FnvHashMap;
+use rustc::util::nodemap::FnvHashMap;
 
 pub const DUMMY_WILD_PAT: &'static Pat = &Pat {
     id: DUMMY_NODE_ID,
@@ -546,7 +546,7 @@ fn construct_witness<'a,'tcx>(cx: &MatchCheckCtxt<'a,'tcx>, ctor: &Constructor,
         ty::TyTuple(_) => PatKind::Tup(pats.collect()),
 
         ty::TyEnum(adt, _) | ty::TyStruct(adt, _)  => {
-            let v = adt.variant_of_ctor(ctor);
+            let v = ctor.variant_for_adt(adt);
             match v.kind() {
                 VariantKind::Struct => {
                     let field_pats: hir::HirVec<_> = v.fields.iter()
@@ -617,13 +617,13 @@ fn construct_witness<'a,'tcx>(cx: &MatchCheckCtxt<'a,'tcx>, ctor: &Constructor,
     })
 }
 
-impl<'tcx, 'container> ty::AdtDefData<'tcx, 'container> {
-    fn variant_of_ctor(&self,
-                       ctor: &Constructor)
-                       -> &VariantDefData<'tcx, 'container> {
-        match ctor {
-            &Variant(vid) => self.variant_with_id(vid),
-            _ => self.struct_variant()
+impl Constructor {
+    fn variant_for_adt<'tcx, 'container, 'a>(&self,
+                                             adt: &'a ty::AdtDefData<'tcx, 'container>)
+                                             -> &'a VariantDefData<'tcx, 'container> {
+        match self {
+            &Variant(vid) => adt.variant_with_id(vid),
+            _ => adt.struct_variant()
         }
     }
 }
@@ -843,7 +843,7 @@ pub fn constructor_arity(_cx: &MatchCheckCtxt, ctor: &Constructor, ty: Ty) -> us
             _ => 1
         },
         ty::TyEnum(adt, _) | ty::TyStruct(adt, _) => {
-            adt.variant_of_ctor(ctor).fields.len()
+            ctor.variant_for_adt(adt).fields.len()
         }
         ty::TyArray(_, n) => n,
         _ => 0
@@ -924,7 +924,7 @@ pub fn specialize<'a>(cx: &MatchCheckCtxt, r: &[&'a Pat],
         PatKind::Struct(_, ref pattern_fields, _) => {
             let def = cx.tcx.def_map.borrow().get(&pat_id).unwrap().full_def();
             let adt = cx.tcx.node_id_to_type(pat_id).ty_adt_def().unwrap();
-            let variant = adt.variant_of_ctor(constructor);
+            let variant = constructor.variant_for_adt(adt);
             let def_variant = adt.variant_of_def(def);
             if variant.did == def_variant.did {
                 Some(variant.fields.iter().map(|sf| {
