+++ src/librustc_lint/Cargo.toml
@@ -12,5 +12,6 @@ crate-type = ["dylib"]
 log = { path = "../liblog" }
 rustc = { path = "../librustc" }
 rustc_back = { path = "../librustc_back" }
+rustc_const_eval = { path = "../librustc_const_eval" }
 rustc_front = { path = "../librustc_front" }
 syntax = { path = "../libsyntax" }
