+++ src/librustc_lint/lib.rs
@@ -46,6 +46,7 @@ extern crate rustc;
 extern crate log;
 extern crate rustc_front;
 extern crate rustc_back;
+extern crate rustc_const_eval;
 
 pub use rustc::lint as lint;
 pub use rustc::middle as middle;
