+++ src/librustc_mir/hair/cx/mod.rs
@@ -18,7 +18,8 @@
 use hair::*;
 use rustc::mir::repr::*;
 
-use rustc::middle::const_eval::{self, ConstVal};
+use rustc::middle::const_val::ConstVal;
+use rustc_const_eval as const_eval;
 use rustc::middle::def_id::DefId;
 use rustc::infer::InferCtxt;
 use rustc::ty::subst::{Subst, Substs};
