+++ src/librustc_mir/lib.rs
@@ -33,6 +33,7 @@ extern crate rustc_front;
 extern crate rustc_back;
 extern crate syntax;
 extern crate rustc_const_math;
+extern crate rustc_const_eval;
 
 pub mod build;
 pub mod graphviz;
