+++ src/librustc_passes/lib.rs
@@ -30,6 +30,7 @@
 extern crate core;
 extern crate rustc;
 extern crate rustc_front;
+extern crate rustc_const_eval;
 
 #[macro_use] extern crate log;
 #[macro_use] extern crate syntax;
