+++ src/librustc_trans/mir/constant.rs
@@ -10,8 +10,9 @@
 
 use llvm::ValueRef;
 use rustc::ty::{Ty, TypeFoldable};
-use rustc::middle::const_eval::{self, ConstVal};
+use rustc::middle::const_val::ConstVal;
 use rustc_const_math::ConstInt::*;
+use rustc_const_eval::lookup_const_by_id;
 use rustc::mir::repr as mir;
 use abi;
 use common::{self, BlockAndBuilder, C_bool, C_bytes, C_floating_f64, C_integral,
@@ -114,7 +115,7 @@ impl<'bcx, 'tcx> MirContext<'bcx, 'tcx> {
                 }
 
                 let substs = Some(bcx.monomorphize(substs));
-                let expr = const_eval::lookup_const_by_id(bcx.tcx(), def_id, substs)
+                let expr = lookup_const_by_id(bcx.tcx(), def_id, substs)
                             .expect("def was const, but lookup_const_by_id failed").0;
                 // FIXME: this is falling back to translating from HIR. This is not easy to fix,
                 // because we would have somehow adapt const_eval to work on MIR rather than HIR.
