+++ src/librustc_trans/mir/rvalue.rs
@@ -11,7 +11,7 @@
 use llvm::ValueRef;
 use rustc::ty::{self, Ty};
 use rustc::ty::cast::{CastTy, IntTy};
-use middle::const_eval::ConstVal;
+use middle::const_val::ConstVal;
 use rustc_const_math::ConstInt;
 use rustc::mir::repr as mir;
