+++ src/librustdoc/Cargo.toml
@@ -14,6 +14,7 @@ arena = { path = "../libarena" }
 getopts = { path = "../libgetopts" }
 rustc = { path = "../librustc" }
 rustc_back = { path = "../librustc_back" }
+rustc_const_eval = { path = "../librustc_const_eval" }
 rustc_driver = { path = "../librustc_driver" }
 rustc_front = { path = "../librustc_front" }
 rustc_lint = { path = "../librustc_lint" }
